Move redirect to front-end only.
